requirements.txt
torch==2.6.0
torchvision==0.21.0  
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.25.2
tqdm==4.66.1
matplotlib==3.7.3
scikit-image==0.24.0
scipy==1.11.4
requests==2.32.3
filelock==3.16.1
fire==0.7.1
humanize==4.11.0
ftfy==6.3.1
regex==2025.7.34
google-cloud-storage==2.10.0


